backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
# HTTP Client & HTML Parsing (CBIC scraper)
httpx>=0.27.0
beautifulsoup4>=4.12.0
# RAG & Embeddings
sentence-transformers==2.3.1
faiss-cpu>=1.7.3
rank-bm25==0.2.2
numpy<2
# OpenAI (for embeddings & LLM)
openai>=1.0.0
# LangGraph & LLM
langgraph>=0.0.40
langchain>=0.1.0
langchain-groq>=0.1.0
# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# PDF Processing (for parsing agent)
pypdf>=3.0.0
PyMuPDF>=1.23.0
# LLM Clients
google-generativeai>=0.8.0
groq>=0.4.0
# OCR (local, privacy-safe)
surya-ocr>=0.8.0
Pillow>=10.0.0
paddleocr>=2.7.0
pytesseract>=0.3.10
# Prompts
jinja2>=3.0.0
# PII Redaction (Presidio NER-based)
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Testing
pytest>=9.0.0

# Async DB drivers
aiosqlite>=0.20.0
asyncpg>=0.29.0

# Migrations
alembic>=1.13.0


# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0

# Redis
redis>=5.0.0
